% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMDMeasure-package.R
\docType{package}
\name{CMDMeasure-package}
\alias{CMDMeasure-package}
\alias{CMDMeasure}
\title{Conditional Mean Dependence Measures via Energy Statistics}
\description{
CMDMeasure: A package for mutual dependence measures via energy statistics
}
\details{
The CMDMeasure package provides measures of conditional mean dependence and tests of conditional mean 
independence.
}
\section{Measuring conditional mean dependence}{

The conditional mean dependence measures include: 
\itemize{
  \item asymmetric measure \eqn{\mathcal{R}_n} based on distance covariance \eqn{\mathcal{V}_n}
  \item symmetric measure \eqn{\mathcal{S}_n} based on distance covariance \eqn{\mathcal{V}_n}
  \item complete measure \eqn{\mathcal{Q}_n} based on complete V-statistics
  \item simplified complete measure \eqn{\mathcal{Q}_n^\star} based on incomplete V-statistics
  \item asymmetric measure \eqn{\mathcal{J}_n} based on complete measure \eqn{\mathcal{Q}_n}
  \item simplified asymmetric measure \eqn{\mathcal{J}_n^\star} based on simplified complete measure 
    \eqn{\mathcal{Q}_n^\star}
  \item symmetric measure \eqn{\mathcal{I}_n} based on complete measure \eqn{\mathcal{Q}_n}
  \item simplified symmetric measure \eqn{\mathcal{I}_n^\star} based on simplified complete measure 
    \eqn{\mathcal{Q}_n^\star}
}
}

\section{Testing conditional mean independence}{

The conditional mean independence tests based on the conditional mean dependence measures are 
implemented as permutation tests.
}

\author{
Ze Jin \email{zj58@cornell.edu},
  Shun Yao \email{shunyao2@illinois.edu},
  Xiaofeng Shao \email{xshao@illinois.edu},
  David S. Matteson \email{matteson@cornell.edu}
}
